fix: resolve clang-tidy warnings for init-variables, format, implicit-conv, isolate-decl
Auto-fixed with clang-tidy --fix across 40 source files:
- cppcoreguidelines-init-variables: add missing initializers
- clang-diagnostic-format: correct printf format specifiers
- readability-implicit-bool-conversion: explicit casts
- readability-isolate-declaration: split multi-declarations

+++ Bootloader/Hal/board_up.c
@@ -21,7 +21,7 @@
 
 bool driverWaitTrueF(bool (*f)(), TickType_t timeout)
 {
-    while(((*f)()) == false)
+    while((int)((*f)()) == false)
     {
         //vTaskDelay(1);
         portDelayus(1);
@@ -40,7 +40,7 @@ bool driverWaitTrueF(bool (*f)(), TickType_t timeout)
 
 bool driverWaitFalseF(bool (*f)(), TickType_t timeout)
 {
-    while(((*f)()) == true)
+    while((int)((*f)()) == true)
     {
         //vTaskDelay(1);
         portDelayus(1);
